Client error handling and video improvements by JT
What I added:

I can think of a couple of ways the client can get stuck on the initial "retrieving" page:  The server isn't running; the wrong address (ip and/or port) was specified for the server; the wrong path was setup for the media on the server; or, the path points to a valid directory but there's no media files present in that directory.

The server address issues I've tried to take care of on the client end.  If the client can't connect to the server the option to enter the address manually is presented to the user or the user can start up the server and try again.  The server address gets stored in the Roku registry and takes precedence over the path hardcoded in the source.

There's not much that can be done about the media path issue from the client side other than to let the user know that there's a problem, so that's what I've done by presenting a message dialog to the user if a connection to the server was made but no items were retrieved.

If anyone knows of another way that the client can get stuck, let me know.

My feature:

I added the ability to resume a video where the user left off.  I accomplished this by having the Roku DVP report its position in the video every 10 seconds.  If the user stops the video before the end is reached, I save the last reported in the registry under a key whose name matches the title of the video.  When a video is played, the registry is checked for the key corresponding to the title of the video and if it exists the user is prompted to resume or start at the beginning.  If/when the video is played the whole way through the key is deleted.

I think it works pretty well, but I can think of one potential problem.  The documentation for the Roku registry isn't so great.  It states each "suite" (whatever that means) is limited to 16K.  16K what?  Bytes I would assume.  I doubt it will be an issue, but if the user is in the habit of playing a video partially and then deleting it off the server without ever playing it the whole way through, there will be orphaned registry keys left on the Roky DVP.  Like I said, I doubt it will be a problem but what does everyone else think?  As far as I could tell there isn't even anyway to tell how full the registry is.

A final note about the registry:  there's a "Transient" section of the registry that I'm using for everything for now.  What that means is that whenever the Roku DVP is rebooted, all the registry entries that I've stored will be lost.  So if you have your Roku DVP connected to a power strip that you turn off regularly you'll lose anything I've stored.  When I'm satisfied everything is working properly I'll move it to the nonvolatile registry.

-JT